puzzles 2 and 3

+++ src/DeterministicDeployFactory.sol
@@ -2,22 +2,17 @@
 pragma solidity ^0.8.9;
 
 contract DeterministicDeployFactory {
-// This is the older way of doing it using assembly
-    event Deployed(address addr, uint salt);
+    // This is the older way of doing it using assembly
+    event Deployed(address addr, uint256 salt);
 
-    function getAddress(
-        bytes memory bytecode,
-        uint _salt
-    ) public view returns (address) {
-        bytes32 hash = keccak256(
-            abi.encodePacked(bytes1(0xff), address(this), _salt, keccak256(bytecode))
-        );
+    function getAddress(bytes memory bytecode, uint256 _salt) public view returns (address) {
+        bytes32 hash = keccak256(abi.encodePacked(bytes1(0xff), address(this), _salt, keccak256(bytecode)));
 
         // NOTE: cast last 20 bytes of hash to address
-        return address(uint160(uint(hash)));
+        return address(uint160(uint256(hash)));
     }
 
-    function deploy(bytes memory bytecode, uint _salt) public payable returns(address) {
+    function deploy(bytes memory bytecode, uint256 _salt) public payable returns (address) {
         address addr;
 
         /*
@@ -31,17 +26,16 @@ contract DeterministicDeployFactory {
               s = big-endian 256-bit value
         */
         assembly {
-            addr := create2(
-                callvalue(), // wei sent with current call
-                // Actual code starts after skipping the first 32 bytes
-                add(bytecode, 0x20),
-                mload(bytecode), // Load the size of code contained in the first 32 bytes
-                _salt // Salt from function arguments
-            )
+            addr :=
+                create2(
+                    callvalue(), // wei sent with current call
+                    // Actual code starts after skipping the first 32 bytes
+                    add(bytecode, 0x20),
+                    mload(bytecode), // Load the size of code contained in the first 32 bytes
+                    _salt // Salt from function arguments
+                )
 
-            if iszero(extcodesize(addr)) {
-                revert(0, 0)
-            }
+            if iszero(extcodesize(addr)) { revert(0, 0) }
         }
 
         emit Deployed(addr, _salt);
